Re-organize SG sorting integration tests
